feat(trusty-search): bundle trusty-embedderd into single-install surface
`cargo install trusty-search` now produces both binaries (trusty-search
and trusty-embedderd) from a single command. Previously users had to run
two separate install commands after the Phase 2 sidecar architecture
(#187) made trusty-embedderd a required runtime dependency.

Changes:
- trusty-embedderd: add [lib] target exposing `pub async fn run()` so the
  daemon's startup logic is callable from downstream crates without any
  duplication. src/main.rs becomes a 3-line shim calling `run().await`.
- trusty-search: add [[bin]] trusty-embedderd (src/bin/trusty-embedderd.rs)
  delegating to `trusty_embedderd::run()`, plus the dep on the library
  crate. Cargo's binary-per-crate rule means `cargo install trusty-search`
  now installs both executables atomically.
- Root Cargo.toml: add trusty-embedderd to [workspace.dependencies] so the
  dep resolves to the in-tree path during development.
- Supervisor discovery: no change needed — `locate_embedderd_binary()` in
  trusty-common already checks `current_exe().parent()` (step 2 in the
  lookup chain), which covers the bundled-install case where both binaries
  land in the same ~/.cargo/bin/ prefix.
- Tests: add `tests/bundled_install.rs` (#[ignore]) that shells out to
  `cargo install --root <tempdir>` and asserts both binaries are present
  and executable.
- Docs: update CHANGELOG.md, CLAUDE.md, and README.md to reflect the
  one-command install; remove "install trusty-embedderd separately" wording.

No behavior change for users who already had trusty-embedderd on PATH.
The upgrade path for Phase 2 users: `cargo install trusty-search --locked --force`.
