+++ crates/trusty-search/CLAUDE.md
@@ -507,7 +507,7 @@ environment variable (issue #110 Phase 2):
 
 | `TRUSTY_EMBEDDER` value | Behaviour |
 |-------------------------|-----------|
-| unset / `auto` / `stdio` | **Default.** Spawns `trusty-embedderd --stdio` as a supervised child process. `trusty-embedderd` is a **required runtime dependency** — if it is not found on PATH and `TRUSTY_EMBEDDERD_BIN` is unset, `trusty-search start` exits with an actionable error. Install with `cargo install trusty-embedderd --locked`. |
+| unset / `auto` / `stdio` | **Default.** Spawns `trusty-embedderd --stdio` as a supervised child process. `trusty-embedderd` is a **required runtime dependency** — if it is not found on PATH and `TRUSTY_EMBEDDERD_BIN` is unset, `trusty-search start` exits with an actionable error. **`cargo install trusty-search` installs `trusty-embedderd` automatically** — no separate install needed. |
 | `in-process` / `local`  | Explicit escape hatch — in-process ONNX embedding. Use for tests, debugging, or environments where the sidecar cannot be installed. **Never activated silently**: you must set this variable explicitly to use the in-process path. |
 | `http://…`              | HTTP remote — `POST /embed` to a manually-managed `trusty-embedderd` HTTP listener. |
 | `unix:/path/to/sock`    | UDS remote — JSON-RPC 2.0 to a manually-managed `trusty-embedderd --socket` listener. |
